+++ docs/app/react/admonition/components/admonition-preview.tsx
@@ -1,122 +1,12 @@
-import { Information } from '@cerberus-design/icons'
-import { Avatar, Tag } from '@cerberus-design/react'
-import { HStack, VStack } from '@cerberus-design/styled-system/jsx'
-import { css, cx } from '@cerberus/styled-system/css'
 import {
-  admonition,
-  type AdmonitionVariantProps,
-} from '@cerberus/styled-system/recipes'
-import type { HTMLAttributes } from 'react'
-
-type MatchAvatarProps = AdmonitionVariantProps
-
-function MatchAvatar(props: MatchAvatarProps) {
-  switch (props.palette) {
-    case 'page':
-      return (
-        <Avatar
-          gradient="charon-light"
-          ariaLabel=""
-          icon={<Information />}
-          size="sm"
-          src=""
-        />
-      )
-    case 'info':
-      return (
-        <Avatar
-          gradient="amphiaraus-dark"
-          ariaLabel=""
-          icon={<Information />}
-          size="sm"
-          src=""
-        />
-      )
-    case 'success':
-      return (
-        <Avatar
-          gradient="thanatos-dark"
-          ariaLabel=""
-          icon={<Information />}
-          size="sm"
-          src=""
-        />
-      )
-    case 'warning':
-      return (
-        <Avatar
-          gradient="asphodel-light"
-          ariaLabel=""
-          icon={<Information />}
-          size="sm"
-          src=""
-        />
-      )
-    case 'danger':
-      return (
-        <Avatar
-          gradient="hades-light"
-          ariaLabel=""
-          icon={<Information />}
-          size="sm"
-          src=""
-        />
-      )
-
-    default:
-      throw new Error('Unsupported admonition palette')
-  }
-}
-
-export type AdmonitionProps = HTMLAttributes<HTMLDivElement> &
-  AdmonitionVariantProps
-
-export function Admonition(props: AdmonitionProps) {
-  const { children, palette = 'page', usage, ...nativeProps } = props
-  return (
-    <HStack
-      {...nativeProps}
-      className={cx(nativeProps.className, admonition({ palette, usage }).root)}
-      gap="md"
-      w="full"
-    >
-      <MatchAvatar palette={palette} />
-      <div>{children}</div>
-    </HStack>
-  )
-}
-
-type AdmonitionHeadingProps = HTMLAttributes<HTMLParagraphElement> &
-  AdmonitionVariantProps
-
-export function AdmonitionHeading(props: AdmonitionHeadingProps) {
-  const { palette, usage, ...nativeProps } = props
-  return (
-    <p
-      {...nativeProps}
-      className={cx(
-        nativeProps.className,
-        admonition({ palette, usage }).heading,
-      )}
-    />
-  )
-}
-
-type AdmonitionDescriptionProps = HTMLAttributes<HTMLParagraphElement> &
-  AdmonitionVariantProps
-
-export function AdmonitionDescription(props: AdmonitionDescriptionProps) {
-  const { palette, usage, ...nativeProps } = props
-  return (
-    <p
-      {...nativeProps}
-      className={cx(
-        nativeProps.className,
-        admonition({ palette, usage }).description,
-      )}
-    />
-  )
-}
+  Admonition,
+  AdmonitionDescription,
+  AdmonitionHeading,
+  Tag,
+  type AdmonitionProps,
+} from '@cerberus-design/react'
+import { VStack } from '@cerberus-design/styled-system/jsx'
+import { css } from '@cerberus/styled-system/css'
 
 function PageAdmonition() {
   const styleProps = {
